fix: logout 디렉토리변경(member->auth)
